Move changeLanguage to its respective fragment
