+++ app/src/main/java/com/delacrixmorgan/kingscup/menu/MenuLanguageFragment.kt
@@ -6,6 +6,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.delacrixmorgan.kingscup.R
+import com.delacrixmorgan.kingscup.common.PreferenceHelper
+import com.delacrixmorgan.kingscup.common.PreferenceHelper.get
+import com.delacrixmorgan.kingscup.common.PreferenceHelper.set
+import com.delacrixmorgan.kingscup.common.setLocale
+import com.delacrixmorgan.kingscup.model.LanguageType
 
 /**
  * MenuLanguageFragment
@@ -18,5 +23,23 @@ import com.delacrixmorgan.kingscup.R
 class MenuLanguageFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_menu_language, container, false)
+    }
+
+    private fun changeLanguage() {
+        val context = this.context ?: return
+        val preference = PreferenceHelper.getPreference(context)
+        val languageTypes = LanguageType.values()
+        var currentLanguage: LanguageType = languageTypes.first {
+            it.countryIso == preference[PreferenceHelper.LANGUAGE, PreferenceHelper.LANGUAGE_DEFAULT]
+        }
+
+        currentLanguage = if (currentLanguage == languageTypes.last()) {
+            languageTypes.first()
+        } else {
+            languageTypes[currentLanguage.ordinal + 1]
+        }
+
+        preference[PreferenceHelper.LANGUAGE] = currentLanguage.countryIso
+        this.resources.setLocale(currentLanguage.countryIso)
     }
 }
