+++ app/src/main/java/com/delacrixmorgan/kingscup/menu/MenuSettingFragment.kt
@@ -7,12 +7,7 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.navigation.Navigation
 import com.delacrixmorgan.kingscup.R
-import com.delacrixmorgan.kingscup.common.PreferenceHelper
-import com.delacrixmorgan.kingscup.common.PreferenceHelper.get
-import com.delacrixmorgan.kingscup.common.PreferenceHelper.set
 import com.delacrixmorgan.kingscup.common.launchShareGameIntent
-import com.delacrixmorgan.kingscup.common.setLocale
-import com.delacrixmorgan.kingscup.model.LanguageType
 import kotlinx.android.synthetic.main.fragment_menu_setting.*
 
 /**
@@ -41,46 +36,19 @@ class MenuSettingFragment : Fragment() {
             Navigation.findNavController(view).navigate(action)
         }
 
-        this.shareViewGroup.setOnClickListener {
-            val message = getString(R.string.preference_message_share_friend)
-            launchShareGameIntent(message)
-        }
-
         this.creditViewGroup.setOnClickListener {
             val action = MenuSettingFragmentDirections.actionMenuSettingFragmentToMenuCreditsFragment()
             Navigation.findNavController(view).navigate(action)
         }
 
         this.languageButton.setOnClickListener {
-            changeLanguage()
-        }
-    }
-
-    private fun changeLanguage() {
-        val context = this.context ?: return
-        val preference = PreferenceHelper.getPreference(context)
-        val languageTypes = LanguageType.values()
-        var currentLanguage: LanguageType = languageTypes.first {
-            it.countryIso == preference[PreferenceHelper.LANGUAGE, PreferenceHelper.LANGUAGE_DEFAULT]
+            val action = MenuSettingFragmentDirections.actionMenuSettingFragmentToMenuLanguageFragment()
+            Navigation.findNavController(view).navigate(action)
         }
 
-        currentLanguage = if (currentLanguage == languageTypes.last()) {
-            languageTypes.first()
-        } else {
-            languageTypes[currentLanguage.ordinal + 1]
+        this.shareViewGroup.setOnClickListener {
+            val message = getString(R.string.preference_message_share_friend)
+            launchShareGameIntent(message)
         }
-
-        preference[PreferenceHelper.LANGUAGE] = currentLanguage.countryIso
-        this.resources.setLocale(currentLanguage.countryIso)
-
-        this.updateSettingLanguage()
-    }
-
-    private fun updateSettingLanguage() {
-        this.settingTextView.text = getString(R.string.preference_title)
-        this.quickGuideTextView.text = getString(R.string.preference_title_quick_guide)
-        this.creditTextView.text = getString(R.string.preference_title_credits_summary)
-        this.shareTextView.text = getString(R.string.preference_title_share_friend)
-        this.languageButton.text = getString(R.string.preference_current_language)
     }
 }
